fixed alarm, formatted things
alarm wasn't ringing as the intent brings it directly back to home before it has a chance to ring. i added a for loop that repeats the tone, hoping that it'll stop it after 5 times, but didn't work so the "in transit" button snoozes the alarm.
didn't delete any files bc not quite sure which ones are needed for maps. deleted comment chunks and unnecessary content.